ai_agents_practical_guides/pyproject.toml
[project]
name = "ai-agents-practical-guides"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "langchain>=1.2.8",
    "langchain-groq>=1.1.2",
    "python-dotenv>=1.2.1",
    "rich>=14.3.2",
]
